Use target spans when checking for "already seen" duplicates
We want to use the target spans rather than the source spans because
there may be several places in the target documents copy pasted from
the same place in the source document, and we want to know about all
of them.

Before this fix there was a consistency issue: target spans were added
to the "seen" set but source spans were used when testing if seen.

Note that we now also use target spans when checking for length of
duplicate because target and source lengths should be the same anyway
so it is more consistent.